Move PipelineJobAPI and RuleAlteredJobAPI from api module to core module
